spottyd/requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.21
argh==0.26.2
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.1
discord.py==1.7.3
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
protobuf==3.17.3
PyMySQL==0.9.3
python-dotenv==0.19.0
PyYAML==5.4.1
six==1.16.0
typing-extensions==3.10.0.0
watchdog==2.1.3
Werkzeug==2.0.1
yarl==1.6.3
